+++ staging_service/AutoDetectUtils.py
@@ -21,10 +21,12 @@ class AutoDetectUtils:
             # preferentially choose the most specific suffix (e.g. longest)
             # to get file type mappings
             for i in range(1, dotcount + 1):
-                suffix = filename.split(".", i)[-1].lower()
+                parts = filename.split(".", i)
+                suffix = parts[-1].lower()
                 if suffix in AutoDetectUtils._MAPPINGS["types"]:
-                    return AutoDetectUtils._MAPPINGS["types"][suffix]
-        return None
+                    prefix = ".".join(parts[0:i])
+                    return AutoDetectUtils._MAPPINGS["types"][suffix], prefix, parts[-1]
+        return None, filename, None
 
     @staticmethod
     def get_mappings(file_list: list) -> dict:
@@ -34,7 +36,14 @@ class AutoDetectUtils:
         :return: return a listing of apps and a listing of extension_mappings for each filename
         """
         mappings = []
+        fileinfo = []
         for filename in file_list:
-            mappings.append(AutoDetectUtils.determine_possible_importers(filename))
-        rv = {"apps": AutoDetectUtils._MAPPINGS["apps"], "mappings": mappings}
+            typemaps, prefix, suffix = AutoDetectUtils.determine_possible_importers(filename)
+            mappings.append(typemaps)
+            fileinfo.append({"prefix": prefix, "suffix": suffix})
+        rv = {
+            "apps": AutoDetectUtils._MAPPINGS["apps"],
+            "mappings": mappings,
+            "fileinfo": fileinfo,
+        }
         return rv